✨ feat(tools): create server using fastify and hono
