+++ CHANGELOG.md
@@ -11,6 +11,9 @@ Generated by [`auto-changelog`](https://github.com/CookPete/auto-changelog).
 
 ### Merged
 
+- refactor: 19th refactoring [`#208`](https://github.com/rjoydip/templ/pull/208)
+- chore(deps): bump valibot from 0.25.0 to 0.26.0 [`#205`](https://github.com/rjoydip/templ/pull/205)
+- chore(deps-dev): bump knip from 3.13.2 to 4.2.1 [`#207`](https://github.com/rjoydip/templ/pull/207)
 - feat: astro project created [`#206`](https://github.com/rjoydip/templ/pull/206)
 - ✨ feat(none): hono and autocannon benchmark [`#202`](https://github.com/rjoydip/templ/pull/202)
 - 🔨 refactor(none): drop @clack/prompts with consola [`#201`](https://github.com/rjoydip/templ/pull/201)
